+++ clang/docs/ReleaseNotes.rst
@@ -210,6 +210,9 @@ AST Matchers
 clang-format
 ------------
 
+- Adds ``AlignAfterOpenBracketBreak`` sub-options for better control of
+  ``AlignAfterOpenBracket`` with ``AlwaysBreak`` or ``BlockIndent`` modes.
+
 libclang
 --------